+++ examples/notes/src/routes/[...404].tsx
@@ -1,0 +1,5 @@
+import { Navigate } from "@solidjs/router";
+
+export default function NotFound() {
+  return <Navigate href="/" />;
+}