+++ packages/start/config/server-handler.ts
@@ -190,7 +190,11 @@ async function handleSingleFlight(sourceEvent: FetchEvent, result: any) {
   if (result instanceof Response) {
     if (result.headers.has("X-Revalidate"))
       revalidate = result.headers.get("X-Revalidate").split(",");
-    if (result.headers.has("Location")) url = result.headers.get("Location");
+    if (result.headers.has("Location"))
+      url = new URL(
+        result.headers.get("Location"),
+        new URL(sourceEvent.request.url).origin + import.meta.env.SERVER_BASE_URL
+      ).toString();
   }
   const event = cloneEvent(sourceEvent) as PageEvent;
   event.request = new Request(url);